requirements.txt
fastapi
uvicorn
pandas
gradio
streamlit
requests
scikit-learn
matplotlib
groq
python-dotenv
python-multipart